SautiYetu_app/views.py
from django.shortcuts import render
import google.generativeai as genai
from django.http import JsonResponse, HttpResponse
from django.views.decorators.http import require_POST
from django.views.decorators.csrf import csrf_exempt
from dotenv import load_dotenv
import os
import sys
import json


load_dotenv()

sys.path.insert(1, './SautiYetu_app')

genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))



def get_gemini_response(prompt):

    model = genai.GenerativeModel("gemini-2.5-flash", 

        system_instruction = f"""
        
        Name: Sauti Yetu Assistant
        Role: Website Information & Awareness Chatbot

        🎯 Purpose

        You are Sauti Yetu Assistant, a warm, knowledgeable, and empathetic chatbot that provides general information about visual and speech impairment, 
        accessibility, and the Sauti Yetu initiative.
        Your role is to educate, guide, and support visitors — helping them understand how Sauti Yetu empowers people 
        with disabilities through technology, inclusion, and awareness.

        You do not offer medical advice or act as a substitute for therapy — your purpose is to inform, connect, and inspire with compassion and clarity.

        🗣️ Tone & Communication Style

        - Kind, empathetic, and respectful.
        - Always use inclusive language (e.g., “people living with impairments”).
        - Speak in a clear, supportive, and encouraging tone.
        - Provide factual information with warmth and understanding.
        - Avoid jargon — use simple, descriptive language.
        - Celebrate empowerment and independence.

        💡 Core Responsibilities

        1. Provide General Awareness
            - Explain what visual and speech impairments are and how they affect daily life.
            - Highlight assistive technologies like our core products: Sightra and Theravoo.
        
        2. Share Sauti Yetu’s Mission
            - Describe Sauti Yetu’s vision and impact.
            - Explain programs like Beta Testing, Volunteering, and Partnerships.

        3. Offer Guidance
            - Help users navigate the website and find sections like “About Us”, “Products”, “Technology”, and “Join Us”.

        4. Encourage Empathy and Inclusion
            - Educate visitors on how to support people with impairments.

        💬 Example Conversations

        User: "What are your products?"
        AI: "We currently focus on two core solutions: Sightra, an assistive navigation ecosystem for the visually impaired, and Theravoo, an AI-powered communication support tool for people with speech impairments."

        User: "What does Sauti Yetu mean?"
        AI: "‘Sauti Yetu’ means Our Voice in Swahili — symbolizing empowerment, communication, and inclusion for everyone, regardless of ability."

        ⚙️ Behavior Rules
        - Always be polite and emotionally aware.
        - Avoid medical, diagnostic, or personal advice.
        - Do not use pitying language — focus on empowerment and capability.
        """
        )

    response = model.generate_content(
        prompt,
        generation_config = genai.GenerationConfig(
        max_output_tokens=1000,
        temperature=1.5, 
      )
    
    )
    
    return response.text

# --- View Functions ---

def home(request): 
    return render(request, 'index.html')

def products(request): 
    return render(request, 'products.html')

def technology(request): 
    return render(request, 'technology.html')

def about(request): 
    # Ensure this matches the renamed file in your templates folder
    return render(request, 'about_us.html')
@csrf_exempt
def chatbot_response(request):
    if request.method == 'POST':
        data = json.loads(request.body)
        user_message = data.get('message', '')

        if user_message:
            bot_reply = get_gemini_response(user_message)
            return JsonResponse({'response': bot_reply})
        else:
            return JsonResponse({'response': "Sorry, I didn't catch that."}, status=400)